+++ rider-location-service/src/main/java/com/inbobwetrust/controller/RiderLocationController.java
@@ -1,5 +1,7 @@
-package com.inbobwetrust.domain;
+package com.inbobwetrust.controller;
 
+import com.inbobwetrust.domain.RiderLocation;
+import com.inbobwetrust.service.RiderLocationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Mono;
